requirements.txt
fastapi
uvicorn
pydantic
sqlparse
openai
openenv-core
requests
